results/plaid-isolated/haiku-medium-run8-files/requirements.txt
plaid-python==25.1.0
plaid-link-verify==1.0.0
flask==3.0.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
